[#89] refactor: setting 관련 훅 폴더 구조 정리
