+++ src/hooks/setting/usePrivacyPolicy.ts
