+++ src/hooks/setting/useTermsOfService.ts
